+++ src/ui/modals/info_list_modal.rs
@@ -1,4 +1,7 @@
+use std::time::Duration;
+
 use anyhow::Result;
+use bon::bon;
 use itertools::Itertools;
 use ratatui::{
     Frame,
@@ -15,6 +18,7 @@ use crate::{
     context::AppContext,
     mpd::commands::Song,
     shared::{
+        ext::duration::DurationExt,
         key_event::KeyEvent,
         macros::pop_modal,
         mouse_event::{MouseEvent, MouseEventKind},
@@ -23,17 +27,42 @@ use crate::{
 };
 
 #[derive(Debug)]
-pub struct SongInfoModal {
+pub struct InfoListModal {
     scrolling_state: DirState<TableState>,
     table_area: Rect,
-    song: Song,
+    items: KeyValues,
+    column_widths: &'static [u16],
+    title: &'static str,
+    size: (u16, u16),
 }
 
-impl SongInfoModal {
-    pub fn new(song: Song) -> Self {
+#[derive(Debug)]
+struct KeyValues(Vec<KeyValue>);
+#[derive(Debug)]
+struct KeyValue {
+    key: String,
+    value: String,
+}
+
+#[bon]
+impl InfoListModal {
+    #[builder]
+    pub fn new(
+        items: impl Into<KeyValues>,
+        title: &'static str,
+        column_widths: &'static [u16],
+        size: Option<(u16, u16)>,
+    ) -> Self {
         let mut scrolling_state = DirState::default();
         scrolling_state.select(Some(0), 0);
-        Self { scrolling_state, song, table_area: Rect::default() }
+        Self {
+            scrolling_state,
+            items: items.into(),
+            table_area: Rect::default(),
+            title,
+            column_widths,
+            size: size.unwrap_or((80, 80)),
+        }
     }
 
     #[allow(clippy::cast_possible_truncation)]
@@ -60,9 +89,9 @@ impl SongInfoModal {
     }
 }
 
-impl Modal for SongInfoModal {
+impl Modal for InfoListModal {
     fn render(&mut self, frame: &mut Frame, app: &mut AppContext) -> Result<()> {
-        let popup_area = frame.area().centered(80, 80);
+        let popup_area = frame.area().centered(self.size.0, self.size.1);
         frame.render_widget(Clear, popup_area);
         if let Some(bg_color) = app.config.theme.modal_background_color {
             frame.render_widget(Block::default().style(Style::default().bg(bg_color)), popup_area);
@@ -73,9 +102,8 @@ impl Modal for SongInfoModal {
             .border_set(border::ROUNDED)
             .border_style(app.config.as_border_style())
             .title_alignment(ratatui::prelude::Alignment::Center)
-            .title("Song info");
+            .title(self.title);
 
-        let (key_col_width, val_col_width) = (30, 70);
         let margin = Margin { horizontal: 1, vertical: 0 };
         let [header_area, table_area] =
             Layout::vertical([Constraint::Length(2), Constraint::Percentage(100)])
@@ -83,86 +111,39 @@ impl Modal for SongInfoModal {
         let header_area = header_area.inner(margin);
         let table_area = table_area.inner(margin);
 
-        let [tag_area, mut value_area] = Layout::horizontal([
-            Constraint::Percentage(key_col_width),
-            Constraint::Percentage(val_col_width),
-        ])
-        .areas(table_area);
-        value_area.width = value_area.width.saturating_sub(1); // account for the column spacing
+        let column_constraints =
+            self.column_widths.iter().map(|w| Constraint::Percentage(*w)).collect_vec();
+        let column_areas = Layout::horizontal(&column_constraints).spacing(1).split(table_area);
 
-        let Self { song, .. } = self;
-        let mut rows = Vec::new();
-
-        rows.extend(SongInfoModal::row("File", tag_area.width, &song.file, value_area.width));
-        let file_name = song.file_name().unwrap_or_default();
-        if !file_name.is_empty() {
-            rows.extend(SongInfoModal::row(
-                "Filename",
-                tag_area.width,
-                &file_name,
-                value_area.width,
-            ));
-        }
-
-        if let Some(title) = song.metadata.get("title") {
-            rows.extend(title.iter().flat_map(|item| {
-                SongInfoModal::row("Title", tag_area.width, item, value_area.width)
-            }));
-        }
-
-        if let Some(artist) = song.metadata.get("artist") {
-            rows.extend(artist.iter().flat_map(|item| {
-                SongInfoModal::row("Artist", tag_area.width, item, value_area.width)
-            }));
-        }
-
-        if let Some(album) = song.metadata.get("album") {
-            rows.extend(album.iter().flat_map(|item| {
-                SongInfoModal::row("Album", tag_area.width, item, value_area.width)
-            }));
-        }
-
-        let duration = song.duration.as_ref().map(|d| d.as_secs().to_string()).unwrap_or_default();
-        if !duration.is_empty() {
-            rows.extend(SongInfoModal::row(
-                "Duration",
-                tag_area.width,
-                &duration,
-                value_area.width,
-            ));
-        }
-
-        rows.extend(
-            song.metadata
-                .iter()
-                .filter(|(key, _)| {
-                    !["title", "album", "artist", "duration"].contains(&(*key).as_str())
-                })
-                .flat_map(|(k, v)| {
-                    v.iter().flat_map(|item| {
-                        SongInfoModal::row(k, tag_area.width, item, value_area.width)
-                    })
-                }),
-        );
+        let rows = self
+            .items
+            .0
+            .iter()
+            .flat_map(|item| {
+                InfoListModal::row(
+                    &item.key,
+                    column_areas[0].width,
+                    &item.value,
+                    column_areas[1].width,
+                )
+            })
+            .collect_vec();
 
         self.scrolling_state.set_content_len(Some(rows.len()));
         self.scrolling_state.set_viewport_len(Some(table_area.height.into()));
 
-        let header_table = Table::new(vec![Row::new([Cell::from("Tag"), Cell::from("Value")])], [
-            Constraint::Percentage(key_col_width),
-            Constraint::Percentage(val_col_width),
-        ])
+        let header_table = Table::new(
+            vec![Row::new([Cell::from("Tag"), Cell::from("Value")])],
+            &column_constraints,
+        )
         .column_spacing(1)
         .block(
             Block::default().borders(Borders::BOTTOM).border_style(app.config.as_border_style()),
         );
-        let table = Table::new(rows, [
-            Constraint::Percentage(key_col_width),
-            Constraint::Percentage(val_col_width),
-        ])
-        .column_spacing(1)
-        .style(app.config.as_text_style())
-        .row_highlight_style(app.config.theme.current_item_style);
+        let table = Table::new(rows, &column_constraints)
+            .column_spacing(1)
+            .style(app.config.as_text_style())
+            .row_highlight_style(app.config.theme.current_item_style);
 
         self.table_area = table_area;
 
@@ -251,5 +232,92 @@ impl Modal for SongInfoModal {
         }
 
         Ok(())
+    }
+}
+
+impl From<Vec<Song>> for KeyValues {
+    fn from(value: Vec<Song>) -> Self {
+        let mut result = Vec::new();
+
+        let total_duration: Duration = value.iter().filter_map(|v| v.duration).sum();
+        let total_artists = value
+            .iter()
+            .filter_map(|v| v.metadata.get("artist"))
+            .flat_map(|tag| tag.iter())
+            .unique()
+            .count();
+        let total_albums = value
+            .iter()
+            .filter_map(|v| v.metadata.get("album"))
+            .flat_map(|tag| tag.iter())
+            .unique()
+            .count();
+        let total_genres = value
+            .iter()
+            .filter_map(|v| v.metadata.get("genre"))
+            .flat_map(|tag| tag.iter())
+            .unique()
+            .count();
+
+        result.push(KeyValue { key: "Songs".to_owned(), value: value.len().to_string() });
+        result
+            .push(KeyValue { key: "Total duration".to_owned(), value: total_duration.to_string() });
+        result.push(KeyValue { key: "Artists".to_owned(), value: total_artists.to_string() });
+        result.push(KeyValue { key: "Albums".to_owned(), value: total_albums.to_string() });
+        result.push(KeyValue { key: "Genres".to_owned(), value: total_genres.to_string() });
+        KeyValues(result)
+    }
+}
+
+impl From<&Song> for KeyValues {
+    fn from(song: &Song) -> Self {
+        let mut result = Vec::new();
+        result.push(KeyValue { key: "File".to_owned(), value: song.file.clone() });
+        let file_name = song.file_name().unwrap_or_default();
+        if !file_name.is_empty() {
+            result.push(KeyValue { key: "Filename".to_owned(), value: file_name.into_owned() });
+        }
+
+        if let Some(title) = song.metadata.get("title") {
+            result.extend(
+                title
+                    .iter()
+                    .map(|item| KeyValue { key: "Title".to_owned(), value: item.to_owned() }),
+            );
+        }
+
+        if let Some(artist) = song.metadata.get("artist") {
+            result.extend(
+                artist
+                    .iter()
+                    .map(|item| KeyValue { key: "Artist".to_owned(), value: item.to_owned() }),
+            );
+        }
+
+        if let Some(album) = song.metadata.get("album") {
+            result.extend(
+                album
+                    .iter()
+                    .map(|item| KeyValue { key: "Album".to_owned(), value: item.to_owned() }),
+            );
+        }
+
+        let duration = song.duration.as_ref().map(|d| d.as_secs().to_string()).unwrap_or_default();
+        if !duration.is_empty() {
+            result.push(KeyValue { key: "Duration".to_owned(), value: duration });
+        }
+
+        result.extend(
+            song.metadata
+                .iter()
+                .filter(|(key, _)| {
+                    !["title", "album", "artist", "duration"].contains(&(*key).as_str())
+                })
+                .flat_map(|(k, v)| {
+                    v.iter().map(|item| KeyValue { key: k.to_owned(), value: item.to_owned() })
+                }),
+        );
+
+        KeyValues(result)
     }
 }